HDDS-12520. Move auditparser under debug (#8041)
