+++ hadoop-ozone/tools/src/main/java/org/apache/hadoop/ozone/debug/audit/parser/handler/TemplateCommandHandler.java
@@ -15,13 +15,13 @@
  * limitations under the License.
  */
 
-package org.apache.hadoop.ozone.audit.parser.handler;
+package org.apache.hadoop.ozone.debug.audit.parser.handler;
 
 import java.sql.SQLException;
 import java.util.concurrent.Callable;
 import org.apache.hadoop.hdds.cli.HddsVersionProvider;
-import org.apache.hadoop.ozone.audit.parser.AuditParser;
-import org.apache.hadoop.ozone.audit.parser.common.DatabaseHelper;
+import org.apache.hadoop.ozone.debug.audit.parser.AuditParser;
+import org.apache.hadoop.ozone.debug.audit.parser.common.DatabaseHelper;
 import picocli.CommandLine.Command;
 import picocli.CommandLine.Parameters;
 import picocli.CommandLine.ParentCommand;
@@ -32,8 +32,8 @@ import picocli.CommandLine.ParentCommand;
 @Command(name = "template",
     aliases = "t",
     description = "Execute template query.\n\n" +
-        "To run a template query:\n" +
-        "ozone auditparser <path to db file> template <template>\n\n" +
+        "To run a template query on the audit logs loaded to the database:\n" +
+        "ozone debug auditparser <path to db file> template <template>\n\n" +
         "Following templates are available:\n" +
         "(Template)              (Description)\n" +
         "top5users              : Top 5 users.\n" +
